Update the reader to be profile-aware
This updates the reader to store preferences in profiles and to
load books from a specified account. This also removes all of the
sync-to-server bookmark code that absolutely should not have been
in the reader activities in the first place (it should be in the
controller where all of the other I/O and networking code has always
been). This also replaces the bookmarks system with one that's
profile aware and always goes via the controller to load and save
bookmarks. This provides an extension point to re-implement syncing
to server in the controller in the future.

https://github.com/io7m/android/issues/8
https://github.com/io7m/android/issues/5