Replace the splash screen
This replaces the existing splash screen with one that redirects
to the profile selection screen or directly to the catalog if the
anonymous profile is enabled. It removes the EULA check and display
as this should arguably be handled another way (apart from anything,
agreeing to the EULA might be profile-specific information).

Fix https://github.com/io7m/android/issues/23