Initial profile idle timer
This implements a basic profile idle timer. Activities in the app that
require a profile to be active now inherit from ProfileTimeOutActivity -
this includes the reader. The timer is set to move the user back to
the profile selection screen when 10 minutes have elapsed without any
user interaction. The user is warned one minute before they are logged
out with a friendly dialog. Dismissing the dialog (or interacting with
the screen in any way) will reset the idle timer.

If the anonymous profile is enabled (in other words, if profiles are
disabled), the ProfileTimeOutActivity does nothing and the timer is
not started.