Add a simple "version" view in Settings
This adds a task to Gradle to insert the current Git revision into a
file named "revision.properties" in the shared application assets.
The application then queries this file at runtime and displays the
commit ID in logs and also in a preferences item in Settings. This
lets testers know *exactly* what code they're running.

Fix #33